Delete dead TL04/TL05 modules, relocate live utilities
Deleted:
- build_mechanistic_defense_evasion_features.py (TL04 builder, 506 lines)
- retrain_mechanistic_v1_model.py (TL05 workflow, 690 lines)
- test_track_l_retrain_mechanistic_v1_model.py (dead TL05 tests, 139 lines)

Relocated live functions:
- bootstrap_holdout_metric_cis, load_tg01_lock, load_v1_lock,
  partition_track_c_defense_columns → candidate_replay.py (only caller)
- load_tl11_feature_provenance → build_tl13 (only caller)
- TL04 path constants inlined into TL13

TL13's ensure_tl04_artifacts now fails fast instead of auto-rebuilding
the dead-ended TL04 pipeline.

494 tests pass, 29s (was 499/64s before this branch).
